Prepare for non-AppStore Paywall
