+++ ExcalidrawZ/Subscription/Paywall.swift
@@ -11,7 +11,6 @@ import StoreKit
 import ChocofordUI
 import Shimmer
 
-
 struct Paywall: View {
     @Environment(\.colorScheme) private var colorScheme
     @Environment(\.dismiss) private var dismiss
