+++ ExcalidrawZ/Subscription/PaywallModifier.swift
@@ -12,7 +12,9 @@ import SwiftyAlert
 
 struct PaywallModifier: ViewModifier {
     @Environment(\.scenePhase) private var scenePhase
+//#if APP_STORE
     @EnvironmentObject private var store: Store
+//#endif
     
 #if canImport(AppKit)
     @State private var window: NSWindow?
@@ -23,6 +25,7 @@ struct PaywallModifier: ViewModifier {
 
     func body(content: Content) -> some View {
         content
+//#if APP_STORE
 #if canImport(AppKit)
             .bindWindow($window)
             .onReceive(NotificationCenter.default.publisher(for: NSWindow.didBecomeKeyNotification)) { notification in
@@ -41,9 +44,10 @@ struct PaywallModifier: ViewModifier {
                 Paywall()
                     .swiftyAlert()
             }
+//#endif
     }
     
-    
+//#if APP_STORE
     private func updateProductStatus() {
         DispatchQueue.main.async {
             let now = Date()
@@ -58,4 +62,5 @@ struct PaywallModifier: ViewModifier {
             }
         }
     }
+//#endif
 }
